Video manager

+++ video_manager/manager/utils.py
@@ -43,24 +43,38 @@ def merge_clips(clips: List[str], input_path, out_path: str) -> str:
     :param clips: list
     :return output merged clip: str
     """
-    dt_finish = extract_datetime(clips[-1]) + get_duration(clips[-1])
+    dt_finish = extract_datetime(clips[-1]) + get_duration(os.path.join(input_path, clips[-1]))
     camera_name = extract_name(clips[0])
     output_name = f'{datetime.strftime(extract_datetime(clips[0]), DATE_FORMAT)}' \
                   f'__' \
                   f'{datetime.strftime(dt_finish, DATE_FORMAT)}_{camera_name}.mp4'
     output_path = os.path.join(out_path, output_name)
 
+    if os.path.exists(output_path):
+        return
+
     with open('input.txt', 'w') as f:
         for filename in clips:
             f.write(f"file '{os.path.join(input_path, filename)}'\n")
+    if os.name == 'nt':
+        command = ['C:\\Program Files\\ffmpeg\\ffmpeg.exe',
+                   '-f', 'concat',
+                   '-safe', '0',
+                   '-i', 'input.txt',
+                   '-c', 'copy',
+                   '-y', output_path]
+    else:
+        command = ['ffmpeg',
+                   '-f', 'concat',
+                   '-safe', '0',
+                   '-i', 'input.txt',
+                   '-c', 'copy',
+                   '-y', output_path]
 
-    subprocess.call(['ffmpeg',
-                     '-f', 'concat',
-                     '-safe', '0',
-                     '-i', 'input.txt',
-                     '-c', 'copy',
-                     '-y', output_path])
+    subprocess.call(command)
 
     os.remove('input.txt')
+    for filename in clips:
+        os.remove(os.path.join(input_path, filename))
 
     return output_name
